+++ petkeeper/src/main/java/com/projet/petkeeper/MainActivity.kt
@@ -12,8 +12,8 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
+import com.projet.petkeeper.dashboard.DashboardRootScreen
 import com.projet.petkeeper.navigation.NavBar
-import com.projet.petkeeper.ui.dashboard.DashboardRootScreen
 import com.projet.petkeeper.ui.theme.PetkeeperTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
