+++ petkeeper/src/main/java/com/projet/petkeeper/dashboard/DashboardRootScreen.kt
@@ -1,4 +1,4 @@
-package com.projet.petkeeper.ui.dashboard
+package com.projet.petkeeper.dashboard
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
